requirements.txt
pika==1.1.0
setuptools_scm==5.0.1
pyyaml==5.3.1
